refactor(onboarding): rename Onboarding.module.css to page...
